requirements-app.txt
# ===========================================
# 앱용 패키지 (Gradio + RAG)
# ===========================================

# 환경변수
python-dotenv

# LLM 클라이언트 (vLLM 서버에 연결)
openai>=1.0.0

# UI
gradio>=4.0.0

# LangChain (RAG)
langchain-core>=0.2.0
langchain-text-splitters>=0.2.0
langchain-community>=0.2.0
langchain-huggingface>=0.0.3

# 임베딩 & 벡터 검색
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0

# PDF 처리
PyMuPDF>=1.23.0

# OCR
mistralai>=1.0.0

# 유틸리티
numpy<2.0.0
requests>=2.28.0
colorlog>=6.0.0
tqdm
